Split WebViewAudioManager into several classes
