+++ features/call/impl/src/main/kotlin/io/element/android/features/call/impl/audio/AndroidWebViewAudioBridge.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2025 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+package io.element.android.features.call.impl.audio
+
+import android.webkit.JavascriptInterface
+import timber.log.Timber
+
+/**
+ * This class is used to handle the audio device selection in the WebView.
+ * It listens for the audio device selection event and calls the callback with the selected device ID.
+ * It also listens for the audio playback started event and calls the callback to notify that the audio track is ready.
+ */
+class AndroidWebViewAudioBridge(
+    private val onAudioDeviceSelected: (String) -> Unit,
+    private val onAudioPlaybackStarted: () -> Unit,
+) {
+    @JavascriptInterface
+    fun setOutputDevice(id: String) {
+        Timber.d("Audio device selected in webview, id: $id")
+        onAudioDeviceSelected(id)
+    }
+
+    @JavascriptInterface
+    fun onTrackReady() {
+        // This method can be used to notify the WebView that the audio track is ready
+        // It can be used to start playing audio or to update the UI
+        Timber.d("Audio track is ready")
+        onAudioPlaybackStarted()
+    }
+}
