+++ features/call/impl/src/main/kotlin/io/element/android/features/call/impl/audio/SerializableAudioDevice.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2025 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+package io.element.android.features.call.impl.audio
+
+import android.media.AudioDeviceInfo
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
+
+/**
+ * This class is used to serialize the audio device information to JSON.
+ */
+@Serializable
+data class SerializableAudioDevice(
+    val id: String,
+    val name: String,
+    @Transient val type: Int = 0,
+    // These have to be part of the constructor for the JSON serializer to pick them up
+    val isEarpiece: Boolean = type == AudioDeviceInfo.TYPE_BUILTIN_EARPIECE,
+    val isSpeaker: Boolean = type == AudioDeviceInfo.TYPE_BUILTIN_SPEAKER,
+    val isExternalHeadset: Boolean = type == AudioDeviceInfo.TYPE_BLUETOOTH_SCO,
+) {
+    companion object {
+        fun fromAudioDeviceInfo(audioDeviceInfo: AudioDeviceInfo): SerializableAudioDevice {
+            return SerializableAudioDevice(
+                id = audioDeviceInfo.id.toString(),
+                name = deviceName(
+                    type = audioDeviceInfo.type,
+                    name = audioDeviceInfo.productName
+                ),
+                type = audioDeviceInfo.type,
+            )
+        }
+    }
+}
